comment.  exactly.

+++ src/PLTPlane.cc
@@ -3,6 +3,7 @@
 
 PLTPlane::PLTPlane ()
 {
+  // Make me, I dare you
 }
 
 
@@ -18,6 +19,7 @@ PLTPlane::~PLTPlane ()
 
 void PLTPlane::AddHit (PLTHit* Hit)
 {
+  // Add a hit.. stored as pointers.. I am not the owner...you are
   fHits.push_back(Hit);
   fChannel = Hit->Channel();
   fROC = Hit->ROC();
@@ -27,17 +29,19 @@ void PLTPlane::AddHit (PLTHit* Hit)
 
 float PLTPlane::Charge ()
 {
+  // Compute the charge on the entire plane
   double Sum = 0;
   for (std::vector<PLTHit*>::iterator it = fHits.begin(); it != fHits.end(); ++it) {
     Sum += (*it)->Charge();
   }
-  return Sum;
+  return (float) Sum;
 }
 
 
 
 int PLTPlane::Channel ()
 {
+  // Which channel is this coming from?
   return fChannel;
 }
 
@@ -45,14 +49,23 @@ int PLTPlane::Channel ()
 
 bool PLTPlane::AddClusterFromSeed (PLTHit* Hit)
 {
+  // Add a cluster from a seed using some very basic clustering.  Not perfect, but better one
+  // on the way
+
+  // New cluster
   PLTCluster* Cluster = new PLTCluster();
+
+  // Add the seed
   Cluster->AddHit(Hit);
+
+  // Check all buddy hits within 3x3
   for (size_t i = 0; i != fHits.size(); ++i) {
     if (abs(fHits[i]->Row() - Hit->Row()) == 1 && abs(fHits[i]->Column() - Hit->Column()) == 1) {
       Cluster->AddHit(fHits[i]);
     }
   }
 
+  // Better add it to the list so we don't forget to delete
   fClusters.push_back(Cluster);
 
   return true;
@@ -62,6 +75,8 @@ bool PLTPlane::AddClusterFromSeed (PLTHit* Hit)
 
 bool PLTPlane::IsBiggestHitIn3x3(PLTHit* Hit)
 {
+  // Just check if a hit is the biggest in a 3x3 around itself
+
   for (size_t i = 0; i != fHits.size(); ++i) {
     if (abs(fHits[i]->Row() - Hit->Row()) == 1 && abs(fHits[i]->Column() - Hit->Column()) <= 1 && fHits[i]->Charge() > Hit->Charge()) {
       return false;
@@ -78,11 +93,13 @@ bool PLTPlane::IsBiggestHitIn3x3(PLTHit* Hit)
 
 void PLTPlane::Clusterize ()
 {
+  // Loop over hits and find biggest..then use as seeds..
   for (size_t i = 0; i != fHits.size(); ++i) {
     if (IsBiggestHitIn3x3(fHits[i])) {
       AddClusterFromSeed(fHits[i]);
     }
   }
+
   return;
 }
 
@@ -90,6 +107,8 @@ void PLTPlane::Clusterize ()
 
 TH2F* PLTPlane::DrawHits2D ()
 {
+  // Draw the plane
+
   TString Name = "Plane_Channel_";
   Name += fChannel;
   Name += "_ROC_";
@@ -104,17 +123,20 @@ TH2F* PLTPlane::DrawHits2D ()
 
 size_t PLTPlane::NHits ()
 {
+  // Number of hits
   return fHits.size();
 }
 
 
 PLTHit* PLTPlane::Hit (size_t const i)
 {
+  // get a specific hit
   return fHits[i];
 }
 
 int PLTPlane::ROC ()
 {
+  // Which roc is this
   return fROC;
 }
